requirements.txt
Pillow==2.4.0
cssselect==0.9.1
lxml==3.3.5
pyquery==1.2.8
wsgiref==0.1.2
